pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "infralight"
version = "0.1.0"
description = "Visualisation and editing tool for SaltStack & Terraform"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Piergiorgio Yankh", email = "piergiorgioyankh@gmail.com" },
]
dependencies = [
    "nicegui>=2.0",
    "Jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9",
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-mock>=3.14",
    "pytest-playwright>=0.7",
    "playwright>=1.40",
    "mypy>=1.13",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
]
ignore = [
    "SIM117",  # NiceGUI uses nested `with` for UI composition
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
